% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-fies.R
\name{compute_fies}
\alias{compute_fies}
\title{Title}
\usage{
compute_fies(
  .data,
  .x_cols,
  ...,
  .type = "prevalence",
  .agg_var = "barangay_geo",
  .agg_area_level = NULL,
  .agg_area_length = c(2, 5, 7, 10)
)
}
\arguments{
\item{.agg_area_length}{}
}
\description{
Title
}
