% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-crosstab.R
\name{generate_crosstab}
\alias{generate_crosstab}
\title{Generate cross-tabulation}
\usage{
generate_crosstab(
  .data,
  x,
  ...,
  total_by_col = FALSE,
  include_frequency = TRUE,
  include_proportion = TRUE,
  include_column_total = TRUE,
  include_row_total = TRUE,
  include_subtotal = FALSE,
  include_zero_value = FALSE,
  convert_to_percent = TRUE,
  decimal_precision = NULL,
  label_stub = NULL,
  label_total = "Total",
  label_subtotal = NULL,
  names_separator = "__",
  x_as_stub = FALSE,
  remove_cols_from_group = NULL,
  weights = NULL
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a \code{tibble}), a lazy data frame (e.g. from \code{dbplyr} or \code{dtplyr}), or Arrow data format.}

\item{x}{\strong{Required}. Variable to be used as categories.}

\item{...}{Tidy-select column names.}

\item{total_by_col}{Whether to apply the sum columnwise if \code{TRUE} or rowwise \code{FALSE}. Default is \code{FALSE}}

\item{include_frequency}{Whether to include frequency columns. Default is \code{TRUE}.}

\item{include_proportion}{Whether to include proportion/percentage columns. Default is \code{TRUE}.}

\item{include_column_total}{Whether to include column total. Default is \code{TRUE}.}

\item{include_row_total}{Whether to include row total. Default is \code{TRUE}.}

\item{include_subtotal}{CURRENTLY IGNORE Whether to include subtotal. Default is \code{FALSE}.}

\item{include_zero_value}{Whether to drop columns with zero (0) values}

\item{convert_to_percent}{Whether to format to percent or proportion. Default is \code{TRUE}.}

\item{decimal_precision}{Specify the precision of rounding the percent or proportion.}

\item{label_stub}{stubhead label (first column).}

\item{label_total}{Label for the overall total.}

\item{label_subtotal}{Label for the subtotal.}

\item{names_separator}{Column separator that defines the table hierarchy.}

\item{weights}{Weights to be applied in the aggregation.}
}
\value{
Returns a cross-table of type \code{tibble}
}
\description{
This function extends the functionality of \code{generate_frequency} by allowing you to generate cross-tabulations of two (2) or more categorical variables.
}
