% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-frequency.R
\name{generate_frequency}
\alias{generate_frequency}
\title{Generate frequency table}
\usage{
generate_frequency(
  .data,
  x,
  label_stub = NULL,
  sort_frequency = FALSE,
  sort_by = "desc",
  include_total = TRUE,
  include_cumulative = TRUE,
  include_zero_value = FALSE
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a \code{tibble}), a lazy data frame (e.g. from \code{dbplyr} or \code{dtplyr}), or Arrow data format.}

\item{x}{\strong{Required}. Variable to be used as categories.}

\item{label_stub}{Stubhead label (first column).}

\item{sort_frequency}{Whether to sort the output. If set to \code{TRUE}, the frequency will be sorted in descending order.}

\item{include_total}{Whether to include row total.}

\item{include_cumulative}{Whether to include cumulative frequencies.}

\item{exclude_zero_value}{Whether to drop categories with zero (0) values}
}
\value{
\code{generate_frequency} returns five (5) columns by default if \code{x_group} is not specified. These include (1) categories of \code{x}, (2) frequency of each category, (3) percent to total, (4) cumulative frequency, and (5) cumulative percent to total.
}
\description{
This function allows you to generates a frequency distribution table (marginal table) of a categorical variable \code{x} specified in its second argument.
}
