% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-multiple-response.R
\name{generate_multiple_response}
\alias{generate_multiple_response}
\title{Generate summary table from a multiple response category}
\usage{
generate_multiple_response(
  .data,
  x,
  ...,
  value_to_count = 1,
  include_frequency = TRUE,
  include_proportion = TRUE,
  convert_to_percent = TRUE,
  format_precision = 2,
  label_stub = NULL,
  label_total = NULL,
  names_separator = ">",
  recode = TRUE,
  recode_variable_position = 5,
  clean_name = TRUE
)
}
\arguments{
\item{.data}{A .data frame, .data frame extension (e.g. a tibble), a lazy .data frame (e.g. from dbplyr or dtplyr), or Arrow .data format.}

\item{x}{\strong{Required}. Row variable to be used as categories.}

\item{...}{tidyselect columns. If only one variable is specified, it will be treated as letter-coded response; otherwise, it will be parsed as columns with binary-coded response.}

\item{value_to_count}{Value used as basis for counting the frequencies.}

\item{include_frequency}{Whether to include frequency columns.}

\item{include_proportion}{Whether to include proportion/percentage columns.}

\item{convert_to_percent}{Whether to format to \code{percent} or \code{proportion}.}

\item{format_precision}{Specify the precision of rounding the percent or proportion.}

\item{label_stub}{Stubhead label (first column).}

\item{label_total}{Column name for the total.}

\item{names_separator}{Column separator that defines the table hierarchy.}

\item{recode}{Whether to recode the variable name first.}

\item{recode_variable_position}{Indicate the position of letter of variable that uniquely identifies each variable.}

\item{clean_name}{Whether to output a clean column name.}
}
\value{
Returns a cross-table of type \code{tibble}
}
\description{
Generate summary table from a multiple response category
}
