% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-tab-by-level.R
\name{generate_tab}
\alias{generate_tab}
\title{Title}
\usage{
generate_tab(
  .data,
  .x_cols,
  .y_cols,
  .agg_var = "barangay_geo",
  .agg_area_level = NULL,
  .agg_area_length = c(2, 5, 7, 10),
  .pivot_cols_to_row_vs = NULL
)
}
\arguments{
\item{.pivot_cols_to_row_vs}{}
}
\description{
Title
}
