% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-tab-by-level.R
\name{generate_tab_by_level}
\alias{generate_tab_by_level}
\title{Generate by aggregation level}
\usage{
generate_tab_by_level(
  .data,
  .x_cols,
  ...,
  .agg_var = "barangay_geo",
  .agg_area_level = NULL,
  .agg_area_length = c(2, 5, 7, 10),
  .total_by_col = F,
  .switch_col = F,
  .split_multiple_response = F,
  .split_multiple_response_width = NULL,
  .pivot_cols_to_row = F,
  .pivot_cols_to_row_vs = NULL,
  .flatten = FALSE
)
}
\arguments{
\item{.pivot_cols_to_row_vs}{}
}
\description{
Generate by aggregation level
}
