% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-as-excel.R
\name{save_as_excel}
\alias{save_as_excel}
\title{Save tables into Excel}
\usage{
save_as_excel(
  .list,
  filename = NULL,
  formatted = TRUE,
  overwrite = TRUE,
  title = NULL,
  export_settings = NULL,
  include_table_list = TRUE,
  tab_variable_name = "tab_name",
  options = list(),
  ...
)
}
\arguments{
\item{.list}{list of data frames to be exported. Each element will be written in its own sheet/tab.}

\item{filename}{File name with \code{.xlsx} extension; e.g., \code{'my_workbook.xlsx'}}

\item{formatted}{Whether to apply default formatting/styling.}

\item{overwrite}{Whether to overwrite the existing file.}

\item{title}{Title of the table.}

\item{export_settings}{Use exporting settings to apply custom rendering.}

\item{include_table_list}{(IGNORED) Whether to include a summary of list of tables generated}

\item{tab_variable_name}{Column name to be used as tab/sheet name.}

\item{options}{}

\item{...}{Additional arguments}
}
\value{
Workbook
}
\description{
Save tables into Excel
}
